+++ packages/jest-mock/src/index.ts
@@ -5,9 +5,35 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import {Mocks} from '@jest/types';
-
 type Global = NodeJS.Global; // | Window – add once TS improves typings;
+
+namespace JestMock {
+  export type ModuleMocker = ModuleMockerClass;
+  export type MockFunctionMetadataType =
+    | 'object'
+    | 'array'
+    | 'regexp'
+    | 'function'
+    | 'constant'
+    | 'collection'
+    | 'null'
+    | 'undefined';
+
+  export type MockFunctionMetadata<
+    T,
+    Y extends Array<unknown>,
+    Type = MockFunctionMetadataType
+  > = {
+    ref?: number;
+    members?: {[key: string]: MockFunctionMetadata<T, Y>};
+    mockImpl?: (...args: Y) => T;
+    name?: string;
+    refID?: number;
+    type?: Type;
+    value?: T;
+    length?: number;
+  };
+}
 
 /**
  * Possible types of a MockFunctionResult.
@@ -34,7 +60,7 @@ type MockFunctionResult = {
   value: unknown;
 };
 
-type MockFunctionState<T, Y extends unknown[]> = {
+type MockFunctionState<T, Y extends Array<unknown>> = {
   calls: Array<Y>;
   instances: Array<T>;
   invocationCallOrder: Array<number>;
@@ -55,24 +81,24 @@ type MockFunctionConfig = {
 
 // see https://github.com/Microsoft/TypeScript/issues/25215
 type NonFunctionPropertyNames<T> = {
-  [K in keyof T]: T[K] extends (...args: any[]) => any ? never : K
+  [K in keyof T]: T[K] extends (...args: Array<any>) => any ? never : K
 }[keyof T] &
   string;
 type FunctionPropertyNames<T> = {
-  [K in keyof T]: T[K] extends (...args: any[]) => any ? K : never
+  [K in keyof T]: T[K] extends (...args: Array<any>) => any ? K : never
 }[keyof T] &
   string;
 
-interface Mock<T, Y extends unknown[] = unknown[]>
+interface Mock<T, Y extends Array<unknown> = Array<unknown>>
   extends Function,
     MockInstance<T, Y> {
   new (...args: Y): T;
   (...args: Y): T;
 }
 
-interface SpyInstance<T, Y extends unknown[]> extends MockInstance<T, Y> {}
+interface SpyInstance<T, Y extends Array<unknown>> extends MockInstance<T, Y> {}
 
-interface MockInstance<T, Y extends unknown[]> {
+interface MockInstance<T, Y extends Array<unknown>> {
   _isMockFunction: true;
   _protoImpl: Function;
   getMockName(): string;
@@ -273,7 +299,7 @@ function getObjectType(value: unknown): string {
   return Object.prototype.toString.apply(value).slice(8, -1);
 }
 
-function getType(ref?: unknown): Mocks.MockFunctionMetadataType | null {
+function getType(ref?: unknown): JestMock.MockFunctionMetadataType | null {
   const typeName = getObjectType(ref);
   if (
     typeName === 'Function' ||
@@ -406,7 +432,7 @@ class ModuleMockerClass {
     return Array.from(slots);
   }
 
-  private _ensureMockConfig<T, Y extends unknown[]>(
+  private _ensureMockConfig<T, Y extends Array<unknown>>(
     f: Mock<T, Y>,
   ): MockFunctionConfig {
     let config = this._mockConfigRegistry.get(f);
@@ -417,7 +443,7 @@ class ModuleMockerClass {
     return config;
   }
 
-  private _ensureMockState<T, Y extends unknown[]>(
+  private _ensureMockState<T, Y extends Array<unknown>>(
     f: Mock<T, Y>,
   ): MockFunctionState<T, Y> {
     let state = this._mockState.get(f);
@@ -439,7 +465,10 @@ class ModuleMockerClass {
     };
   }
 
-  private _defaultMockState<T, Y extends unknown[]>(): MockFunctionState<T, Y> {
+  private _defaultMockState<T, Y extends Array<unknown>>(): MockFunctionState<
+    T,
+    Y
+  > {
     return {
       calls: [],
       instances: [],
@@ -448,32 +477,32 @@ class ModuleMockerClass {
     };
   }
 
-  private _makeComponent<T, Y extends unknown[]>(
-    metadata: Mocks.MockFunctionMetadata<T, Y, 'object'>,
+  private _makeComponent<T, Y extends Array<unknown>>(
+    metadata: JestMock.MockFunctionMetadata<T, Y, 'object'>,
     restore?: () => void,
   ): Object;
-  private _makeComponent<T, Y extends unknown[]>(
-    metadata: Mocks.MockFunctionMetadata<T, Y, 'array'>,
+  private _makeComponent<T, Y extends Array<unknown>>(
+    metadata: JestMock.MockFunctionMetadata<T, Y, 'array'>,
     restore?: () => void,
   ): Array<unknown>;
-  private _makeComponent<T, Y extends unknown[]>(
-    metadata: Mocks.MockFunctionMetadata<T, Y, 'regexp'>,
+  private _makeComponent<T, Y extends Array<unknown>>(
+    metadata: JestMock.MockFunctionMetadata<T, Y, 'regexp'>,
     restore?: () => void,
   ): RegExp;
-  private _makeComponent<T, Y extends unknown[]>(
-    metadata: Mocks.MockFunctionMetadata<
+  private _makeComponent<T, Y extends Array<unknown>>(
+    metadata: JestMock.MockFunctionMetadata<
       T,
       Y,
       'constant' | 'collection' | 'null' | 'undefined'
     >,
     restore?: () => void,
   ): T;
-  private _makeComponent<T, Y extends unknown[]>(
-    metadata: Mocks.MockFunctionMetadata<T, Y, 'function'>,
+  private _makeComponent<T, Y extends Array<unknown>>(
+    metadata: JestMock.MockFunctionMetadata<T, Y, 'function'>,
     restore?: () => void,
   ): Mock<T, Y>;
-  private _makeComponent<T, Y extends unknown[]>(
-    metadata: Mocks.MockFunctionMetadata<T, Y>,
+  private _makeComponent<T, Y extends Array<unknown>>(
+    metadata: JestMock.MockFunctionMetadata<T, Y>,
     restore?: () => void,
   ): Object | Array<unknown> | RegExp | T | undefined | Mock<T, Y> {
     if (metadata.type === 'object') {
@@ -718,8 +747,8 @@ class ModuleMockerClass {
     }
   }
 
-  private _createMockFunction<T, Y extends unknown[]>(
-    metadata: Mocks.MockFunctionMetadata<T, Y>,
+  private _createMockFunction<T, Y extends Array<unknown>>(
+    metadata: JestMock.MockFunctionMetadata<T, Y>,
     mockConstructor: Function,
   ): Function {
     let name = metadata.name;
@@ -778,8 +807,8 @@ class ModuleMockerClass {
     return createConstructor(mockConstructor);
   }
 
-  private _generateMock<T, Y extends unknown[]>(
-    metadata: Mocks.MockFunctionMetadata<T, Y>,
+  private _generateMock<T, Y extends Array<unknown>>(
+    metadata: JestMock.MockFunctionMetadata<T, Y>,
     callbacks: Array<Function>,
     refs: {
       [key: string]:
@@ -828,10 +857,10 @@ class ModuleMockerClass {
    * @param _metadata Metadata for the mock in the schema returned by the
    * getMetadata method of this module.
    */
-  generateFromMetadata<T, Y extends unknown[]>(
-    _metadata: Mocks.MockFunctionMetadata<T, Y>,
+  generateFromMetadata<T, Y extends Array<unknown>>(
+    _metadata: JestMock.MockFunctionMetadata<T, Y>,
   ): Mock<T, Y> {
-    const callbacks: Function[] = [];
+    const callbacks: Array<Function> = [];
     const refs = {};
     const mock = this._generateMock(_metadata, callbacks, refs);
     callbacks.forEach(setter => setter());
@@ -842,10 +871,10 @@ class ModuleMockerClass {
    * @see README.md
    * @param component The component for which to retrieve metadata.
    */
-  getMetadata<T, Y extends unknown[]>(
+  getMetadata<T, Y extends Array<unknown>>(
     component: T,
     _refs?: Map<T, number>,
-  ): Mocks.MockFunctionMetadata<T, Y> | null {
+  ): JestMock.MockFunctionMetadata<T, Y> | null {
     const refs = _refs || new Map<T, number>();
     const ref = refs.get(component);
     if (ref != null) {
@@ -857,7 +886,7 @@ class ModuleMockerClass {
       return null;
     }
 
-    const metadata: Mocks.MockFunctionMetadata<T, Y> = {type};
+    const metadata: JestMock.MockFunctionMetadata<T, Y> = {type};
     if (
       type === 'constant' ||
       type === 'collection' ||
@@ -880,7 +909,7 @@ class ModuleMockerClass {
     refs.set(component, metadata.refID);
 
     let members: {
-      [key: string]: Mocks.MockFunctionMetadata<T, Y>;
+      [key: string]: JestMock.MockFunctionMetadata<T, Y>;
     } | null = null;
     // Leave arrays alone
     if (type !== 'array') {
@@ -915,7 +944,9 @@ class ModuleMockerClass {
     return !!fn && fn._isMockFunction === true;
   }
 
-  fn<T, Y extends unknown[]>(implementation?: (...args: Y) => T): Mock<T, Y> {
+  fn<T, Y extends Array<unknown>>(
+    implementation?: (...args: Y) => T,
+  ): Mock<T, Y> {
     const length = implementation ? implementation.length : 0;
     const fn = this._makeComponent<T, Y>({length, type: 'function'});
     if (implementation) {
@@ -939,7 +970,7 @@ class ModuleMockerClass {
   spyOn<T extends {}, M extends FunctionPropertyNames<T>>(
     object: T,
     methodName: M,
-  ): T[M] extends (...args: any[]) => any
+  ): T[M] extends (...args: Array<any>) => any
     ? SpyInstance<ReturnType<T[M]>, Parameters<T[M]>>
     : never;
 
@@ -1077,7 +1108,6 @@ class ModuleMockerClass {
   }
 }
 
-// TODO: bring this type export back once done with TS migration
-// export type ModuleMocker = ModuleMockerClass;
-
-export = new ModuleMockerClass(global);
+/* eslint-disable-next-line no-redeclare */
+const JestMock = new ModuleMockerClass(global);
+export = JestMock;